[add] client side routing and SCSS
