+++ public/stylesheets/main/splash.scss
@@ -3,7 +3,6 @@ body {
    background-image: url('http://s7d4.scene7.com/is/image/roomandboard/hutton_706468_16e_g?$str_g$&size=1520,960&scl=1');
    background-size: 100%;
    background-repeat: no-repeat;
-
 }
 
 .main {
